requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
sqlparse==0.4.2
backports.zoneinfo==0.2.1
psycopg2
gunicorn